+++ Sources/Moya/URLRequest+Encoding.swift
@@ -2,10 +2,10 @@ import Foundation
 
 internal extension URLRequest {
 
-    mutating func encoded(encodable: Encodable) throws -> URLRequest {
+    mutating func encoded(encodable: Encodable, encoder: JSONEncoder = JSONEncoder()) throws -> URLRequest {
         do {
             let encodable = AnyEncodable(encodable)
-            httpBody = try JSONEncoder().encode(encodable)
+            httpBody = try encoder.encode(encodable)
 
             let contentTypeHeaderName = "Content-Type"
             if value(forHTTPHeaderField: contentTypeHeaderName) == nil {